requirements.txt
tensorflow==2.16.2
pillow==10.4.0
numpy==1.26.4
streamlit==1.41.0